+++ kirlent_docutils/__init__.py
@@ -6,3 +6,12 @@
 """Custom writers for docutils."""
 
 __version__ = "0.1.0"
+
+from docutils.core import publish_cmdline
+
+from . import html5
+
+
+def publish_cmdline_html5(*args, **kwargs):
+    """Run utility for converting an RST file to HTML5."""
+    publish_cmdline(*args, **kwargs, writer=html5.Writer())